removed user from adminpanel
